Add error and loading widget to common and move crypto list to separate folder
