+++ lib/view/common/loading.dart
@@ -1,0 +1,29 @@
+import 'package:flutter/material.dart';
+
+class Loading extends StatelessWidget {
+  final String loadingMessage;
+
+  const Loading({Key key, this.loadingMessage}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Center(
+      child: Column(
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: <Widget>[
+          Text(
+            loadingMessage,
+            textAlign: TextAlign.center,
+            style: TextStyle(
+              fontSize: 24,
+            ),
+          ),
+          SizedBox(height: 24),
+          CircularProgressIndicator(
+            valueColor: AlwaysStoppedAnimation<Color>(Colors.lightGreen),
+          ),
+        ],
+      ),
+    );
+  }
+}
