+++ lib/view/crypto/crypto_list.dart
@@ -1,6 +1,8 @@
 import 'package:crypto_reminder/blocs/cryptocurrency_bloc.dart';
 import 'package:crypto_reminder/models/cryptocurrency.dart';
 import 'package:crypto_reminder/networking/api_response.dart';
+import 'package:crypto_reminder/view/common/error.dart';
+import 'package:crypto_reminder/view/common/loading.dart';
 import 'package:flutter/material.dart';
 
 class CryptoScreen extends StatefulWidget {
@@ -80,7 +82,8 @@ class CryptoList extends StatelessWidget {
               scrollDirection: Axis.vertical,
               itemCount: cryptoList.length,
               itemBuilder: (BuildContext ctxt, int index) {
-                return new Text('${cryptoList[index].symbol}: ${cryptoList[index].lastPrice}');
+                return new Text(
+                    '${cryptoList[index].symbol}: ${cryptoList[index].lastPrice}');
               },
             ),
           ),
@@ -91,70 +94,6 @@ class CryptoList extends StatelessWidget {
         ),
       ],
       mainAxisAlignment: MainAxisAlignment.spaceBetween,
-    );
-  }
-}
-
-class Error extends StatelessWidget {
-  final String errorMessage;
-
-  final Function onRetryPressed;
-
-  const Error({Key key, this.errorMessage, this.onRetryPressed})
-      : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Center(
-      child: Column(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: <Widget>[
-          Text(
-            errorMessage,
-            textAlign: TextAlign.center,
-            style: TextStyle(
-              color: Colors.red,
-              fontSize: 18,
-            ),
-          ),
-          SizedBox(height: 8),
-          RaisedButton(
-            color: Colors.redAccent,
-            child: Text(
-              'Retry',
-            ),
-            onPressed: onRetryPressed,
-          )
-        ],
-      ),
-    );
-  }
-}
-
-class Loading extends StatelessWidget {
-  final String loadingMessage;
-
-  const Loading({Key key, this.loadingMessage}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Center(
-      child: Column(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: <Widget>[
-          Text(
-            loadingMessage,
-            textAlign: TextAlign.center,
-            style: TextStyle(
-              fontSize: 24,
-            ),
-          ),
-          SizedBox(height: 24),
-          CircularProgressIndicator(
-            valueColor: AlwaysStoppedAnimation<Color>(Colors.lightGreen),
-          ),
-        ],
-      ),
     );
   }
 }
